+++ packages/components/src/components/radio-group/radio-group-option/radio-group-option.spec.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it, vi } from 'vitest';
 import * as loggerUtils from '../../../utils/log/logger';
-import * as throwIfPropIsUndefinedUtils from '../../../utils/validation/throwIfPropIsUndefined';
 import { RadioGroupOption } from './radio-group-option';
 
 const initComponent = (): RadioGroupOption => {
@@ -11,12 +10,24 @@ const initComponent = (): RadioGroupOption => {
 };
 
 describe('render', () => {
-  it('should validate that value is defined', () => {
-    const spy = vi.spyOn(throwIfPropIsUndefinedUtils, 'throwIfPropIsUndefined');
+  it('should not throw when value is undefined', () => {
     const component = initComponent();
 
-    expect(() => component.render()).toThrowError(/required property 'value' is undefined/);
-    expect(spy).toHaveBeenCalledWith(component.host, 'value', component.value);
+    expect(() => component.render()).not.toThrow();
+  });
+
+  it('should dispatch "internalOptionValueChange" event when value changes', () => {
+    const component = initComponent();
+    const dispatchEventSpy = vi.spyOn(component.host, 'dispatchEvent');
+
+    component.onValueChange();
+
+    expect(dispatchEventSpy).toHaveBeenCalledWith(
+      expect.objectContaining({
+        type: 'internalOptionValueChange',
+        bubbles: true,
+      })
+    );
   });
 
   it('should reject null values', () => {