+++ packages/components/src/components/select/select-option/select-option.spec.ts
@@ -49,5 +49,18 @@ describe('hasValue computation passed to getOptionAriaAttributes', () => {
   });
 });
 
+describe('onValueChange', () => {
+  it('should dispatch "internalOptionValueChange" event', () => {
+    const component = initComponent();
+    const dispatchEventSpy = vi.spyOn(component.host, 'dispatchEvent');
 
+    component.onValueChange();
 
+    expect(dispatchEventSpy).toHaveBeenCalledWith(
+      expect.objectContaining({
+        type: 'internalOptionValueChange',
+        bubbles: true,
+      })
+    );
+  });
+});